Add files via upload
This document contains the architectural design of the project. It has the UML diagrams, context
diagram, data flow diagram, deployment diagram and various decisions that led to choosing the
described softwares for the development of the project. It contains the description of software
which were used why they were used over other prominent softwares for the same task.